Hardcode const uniforms into the shader itself
Some uniforms aren't actually animated, so they can get hardcoded as
consts in the shader.
They can then get minified through the minifier.
It's still required to check if this actually reduces code size,
although I think it will.
To check, a feature flag is added to disable this behavior.