Save camera rotation keyframes instead of target
This is because internally, the camera controller saves the rotation
and then calculates the target from it.
If the target is saved, what happens is that the rotation is calculated
from the target, which is then immediately calculated from the rotation,
and this goes on and loses accuracy and causes weird glitches.