Switch camera rotation uniform for target
A rotation uniform is now passed, the target uniform is also still
passed.
I'm not 100% sure but the idea is to pass the uniform that is actually
going to be interpolated in the release, because interpolating the
target will both look worse and take up more size (rotation doesn't
have roll so it only needs two variables instead of 3).