Change tonemap and gain functions
The previous tonemap function doesn't have a derivative of 0 at 0,
meaning it isn't flat, which causes it to suddenly jump at the start,
making it not quite "easing".
The solution is changing it to a simple power, which is flat at 0.

The gain function isn't changed, but the factor now only allows for
values above 1, because the values below 1 cause an accelerate-in-out
function, which looks bad, nobody needs, screws up the graph, and also
bugs out changing keyframes in-place, probably because it invalidates
the interpolation function (it's not 0 at 0 or 1 at 1).